break out some soon to be repeated functionality in public pack
