+++ app/javascript/lib/renderComponents.js
@@ -1,0 +1,8 @@
+import renderComponentsForSelector from "../renderComponentsForSelector";
+
+const renderComponents = componentMap =>
+  Object.entries(componentMap).forEach(([selector, component]) => {
+    renderComponentsForSelector(selector, component);
+  });
+
+export default renderComponents;
